refactor: split server graph routes
