+++ server/handlers.go
@@ -7,7 +7,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"net/http"
 	"path"
 	"strconv"
@@ -18,12 +17,9 @@ import (
 	"github.com/dewebprotocol/malt/core/graph"
 	"github.com/dewebprotocol/malt/core/layout/malt/unixfs"
 	"github.com/dewebprotocol/malt/core/manifest"
-	"github.com/dewebprotocol/malt/core/querypath"
 	"github.com/dewebprotocol/malt/core/resolver"
 	"github.com/dewebprotocol/malt/core/resolver/step/explicit"
-	"github.com/dewebprotocol/malt/core/structure"
 	"github.com/dewebprotocol/malt/core/structure/list"
-	"github.com/dewebprotocol/malt/core/structure/mapping"
 	"github.com/dewebprotocol/malt/core/types/arcset"
 	"github.com/dewebprotocol/malt/core/types/evidence"
 	"github.com/dewebprotocol/malt/core/types/prooflist"
@@ -39,378 +35,6 @@ type pathResolution struct {
 	target    cid.Cid
 	stat      *httpapi.PathStatResponse
 	proofList *prooflist.ProofList
-}
-
-func (s *Server) handleHealth(w http.ResponseWriter, r *http.Request) {
-	writeJSON(w, http.StatusOK, &httpapi.HealthResponse{Status: "ok"})
-}
-
-func (s *Server) handleMetrics(w http.ResponseWriter, r *http.Request) {
-	writeJSON(w, http.StatusOK, &httpapi.MetricsResponse{Snapshot: s.node.MetricsSnapshot()})
-}
-
-func (s *Server) handleMetricsReset(w http.ResponseWriter, r *http.Request) {
-	s.node.ResetMetrics()
-	writeJSON(w, http.StatusOK, &httpapi.MetricsResponse{Snapshot: s.node.MetricsSnapshot()})
-}
-
-func (s *Server) handleResolve(w http.ResponseWriter, r *http.Request) {
-	g, err := s.getOrCreateGraph(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-	root, err := decodeCID(r.PathValue("root"))
-	if err != nil {
-		writeError(w, http.StatusBadRequest, "invalid root CID: "+err.Error())
-		return
-	}
-	s.serveResolve(w, r, g, root, r.PathValue("path"))
-}
-
-func (s *Server) handleContent(w http.ResponseWriter, r *http.Request) {
-	g, err := s.getOrCreateGraph(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-	root, err := decodeCID(r.PathValue("root"))
-	if err != nil {
-		writeError(w, http.StatusBadRequest, "invalid root CID: "+err.Error())
-		return
-	}
-	path := r.PathValue("path")
-
-	if _, ok := r.URL.Query()["format"]; ok {
-		writeError(w, http.StatusBadRequest, "format query is not supported; use /resolve/{root}/{path} for path resolution or GET /{root}/{path} for content reads")
-		return
-	}
-
-	wantProof := r.Method != http.MethodHead && !shouldOmitDefaultProof(r)
-	resolved, err := s.resolvePath(r.Context(), g, root, path, wantProof)
-	if err != nil {
-		status := http.StatusInternalServerError
-		if errors.Is(err, errPathNotFound) || errors.Is(err, resolver.ErrResolutionFailed) {
-			status = http.StatusNotFound
-		}
-		writeError(w, status, err.Error())
-		return
-	}
-	stat := resolved.stat
-
-	// HEAD request — return stat headers without body
-	if r.Method == http.MethodHead {
-		w.Header().Set("X-Malt-Kind", stat.Kind)
-		w.Header().Set("X-Malt-Storage-Kind", stat.StorageKind)
-		w.Header().Set("X-Malt-Key", stat.Key)
-		if stat.Payload != "" {
-			w.Header().Set("X-Malt-Payload", stat.Payload)
-		}
-		if stat.Size != nil {
-			w.Header().Set("Content-Length", strconv.FormatInt(*stat.Size, 10))
-		}
-		w.WriteHeader(http.StatusOK)
-		return
-	}
-
-	if stat.Kind == "dir" {
-		// Return JSON directory listing
-		if resolved.proofList != nil {
-			if err := writeProofListHeader(w, *resolved.proofList); err != nil {
-				writeError(w, http.StatusInternalServerError, err.Error())
-				return
-			}
-			s.node.RecordProofList(*resolved.proofList)
-		}
-		addVaryHeader(w, "X-Malt-Proof")
-		writeJSON(w, http.StatusOK, stat)
-		return
-	}
-
-	// Serve file content
-	key, err := decodeCID(stat.Key)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-	totalSize := int64(0)
-	if stat.Size != nil {
-		totalSize = *stat.Size
-	}
-	start, endExclusive, partial, err := parseRangeHeader(r.Header.Get("Range"), totalSize)
-	if err != nil {
-		writeError(w, httpapi.StatusRangeNotSatisfiable, err.Error())
-		return
-	}
-	payload, err := s.readContentPayload(r.Context(), g, stat, key, start, endExclusive)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	// Generate and write proof headers before response headers
-	if resolved.proofList != nil {
-		pl, err := s.readProofList(r.Context(), g, resolved, start, endExclusive)
-		if err != nil {
-			writeError(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-		if err := writeProofListHeader(w, *pl); err != nil {
-			writeError(w, http.StatusInternalServerError, err.Error())
-			return
-		}
-		s.node.RecordProofList(*pl)
-	}
-
-	// Advertise X-Malt-Proof as a variance header since proof generation depends on it
-	addVaryHeader(w, "X-Malt-Proof")
-	w.Header().Set("Accept-Ranges", "bytes")
-	if partial {
-		w.Header().Set("Content-Range", fmt.Sprintf("bytes %d-%d/%d", start, endExclusive-1, totalSize))
-		w.WriteHeader(http.StatusPartialContent)
-	} else {
-		w.WriteHeader(http.StatusOK)
-	}
-	_, _ = io.Copy(w, bytes.NewReader(payload))
-}
-
-func (s *Server) serveResolve(w http.ResponseWriter, r *http.Request, g graph.Runtime, root cid.Cid, queryPath string) {
-	resolved, err := s.resolvePath(r.Context(), g, root, queryPath, !shouldOmitDefaultProof(r))
-	if err != nil {
-		status := http.StatusInternalServerError
-		if errors.Is(err, errPathNotFound) || errors.Is(err, resolver.ErrResolutionFailed) {
-			status = http.StatusNotFound
-		}
-		writeError(w, status, err.Error())
-		return
-	}
-	addVaryHeader(w, "X-Malt-Proof")
-	resp := &httpapi.ResolveResponse{Target: resolved.target.String()}
-	if resolved.proofList != nil {
-		resp.ProofList = resolved.proofList
-		s.node.RecordProofList(*resolved.proofList)
-	}
-	writeJSON(w, http.StatusOK, resp)
-}
-
-func (s *Server) handleSemanticMutation(w http.ResponseWriter, r *http.Request) {
-	g, err := s.getOrCreateGraph(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-	baseRoot, err := decodeCID(r.PathValue("root"))
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-
-	var req httpapi.SemanticMutationRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
-		return
-	}
-
-	mut, err := semanticMutationFromRequest(baseRoot, req.Deltas)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-
-	receipt, err := s.applyWriterMutation(r.Context(), g, mut)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-
-	writeJSON(w, http.StatusCreated, &httpapi.SemanticMutationResponse{
-		BaseRoot:        receipt.BaseRoot.String(),
-		NewRoot:         receipt.NewRoot.String(),
-		ResultRoot:      receipt.NewRoot.String(),
-		DeltaCount:      receipt.DeltaCount,
-		ArcCount:        receipt.ArcCount,
-		MALTObjectCount: receipt.DeltaCount,
-		MapCount:        countSemanticDeltas(mut.Deltas, arcset.KindMap),
-		ListCount:       countSemanticDeltas(mut.Deltas, arcset.KindList),
-	})
-}
-
-func (s *Server) handleWrite(w http.ResponseWriter, r *http.Request) {
-	g, err := s.getOrCreateGraph(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	root, err := decodeCID(r.PathValue("root"))
-	if err != nil {
-		writeError(w, http.StatusBadRequest, "invalid root CID: "+err.Error())
-		return
-	}
-
-	p := r.PathValue("path")
-	if p == "" {
-		writeError(w, http.StatusBadRequest, "path is required")
-		return
-	}
-
-	layout, err := s.unixFSLayout(g)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	baseRoot, err := s.prepareUnixFSRoot(r.Context(), g, layout, root)
-	if err != nil {
-		writeError(w, http.StatusConflict, err.Error())
-		return
-	}
-
-	if r.URL.Query().Get("type") == "dir" {
-		newRoot, err := layout.AddDirectory(r.Context(), baseRoot, p)
-		if err != nil {
-			writeError(w, http.StatusBadRequest, err.Error())
-			return
-		}
-		receipt, err := s.applyUnixFSLayoutMutation(r.Context(), g, layout, root, newRoot)
-		if err != nil {
-			writeError(w, http.StatusBadRequest, err.Error())
-			return
-		}
-		resp := &httpapi.UnixFSWriteResponse{
-			Path:     p,
-			Kind:     "dir",
-			NewRoot:  receipt.NewRoot.String(),
-			ArcCount: receipt.ArcCount,
-		}
-		if root.Defined() {
-			resp.OldRoot = root.String()
-		}
-		writeJSON(w, http.StatusCreated, resp)
-		return
-	}
-
-	data, err := io.ReadAll(r.Body)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, fmt.Sprintf("read request body: %v", err))
-		return
-	}
-	newRoot, err := layout.AddFile(r.Context(), baseRoot, p, data)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	receipt, err := s.applyUnixFSLayoutMutation(r.Context(), g, layout, root, newRoot)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	resp := &httpapi.UnixFSWriteResponse{
-		Path:     p,
-		Kind:     "file",
-		NewRoot:  receipt.NewRoot.String(),
-		ArcCount: receipt.ArcCount,
-	}
-	if root.Defined() {
-		resp.OldRoot = root.String()
-	}
-	writeJSON(w, http.StatusCreated, resp)
-}
-
-func (s *Server) handleCreateStructure(w http.ResponseWriter, r *http.Request) {
-	g, err := s.getOrCreateGraph(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	var req httpapi.CreateStructureRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
-		return
-	}
-	if len(req.Arcs) == 0 {
-		writeError(w, http.StatusBadRequest, "arcs is required")
-		return
-	}
-
-	parsedArcs, err := parseArcMap(req.Arcs)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	snapshot, _, err := buildCreateSnapshot(parsedArcs)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	root, err := g.Writer().CreateStructure(r.Context(), g.Namespace(), snapshot)
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-	writeJSON(w, http.StatusCreated, &httpapi.CreateStructureResponse{Root: root.String()})
-}
-
-func (s *Server) handleVerify(w http.ResponseWriter, r *http.Request) {
-	g, err := s.getOrCreateGraph(r.Context())
-	if err != nil {
-		writeError(w, http.StatusInternalServerError, err.Error())
-		return
-	}
-
-	var req httpapi.VerifyRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		writeError(w, http.StatusBadRequest, fmt.Sprintf("invalid JSON: %v", err))
-		return
-	}
-	valid, err := s.verifyProofList(g, req.ProofList)
-	if err != nil {
-		writeError(w, http.StatusBadRequest, err.Error())
-		return
-	}
-	writeJSON(w, http.StatusOK, &httpapi.VerifyResponse{Valid: valid})
-}
-
-func (s *Server) verifyProofList(g graph.Runtime, pl prooflist.ProofList) (bool, error) {
-	if err := pl.ValidateShape(prooflist.RequireSteps()); err != nil {
-		return false, err
-	}
-	var verifiedPath proofListVerifiedPath
-	for i, step := range pl.Steps {
-		ok, err := s.verifyProofListStep(g, i, step)
-		if err != nil || !ok {
-			return ok, err
-		}
-		if err := verifiedPath.addStep(step); err != nil {
-			return false, err
-		}
-	}
-	if err := validateProofListQuery(pl, verifiedPath); err != nil {
-		return false, err
-	}
-	return true, nil
-}
-
-func validateProofListQuery(pl prooflist.ProofList, verifiedPath proofListVerifiedPath) error {
-	want := arcset.CanonicalizePath(querypath.CanonicalizeQueryPath(pl.Query)).String()
-	if want == "" {
-		return nil
-	}
-	got := verifiedPath.logicalQueryPath()
-	if got == want {
-		return nil
-	}
-	if verifiedPath.hasPayloadBinding {
-		payloadQuery := "@payload"
-		if got != "" {
-			payloadQuery = got + "/@payload"
-		}
-		if want == payloadQuery {
-			return nil
-		}
-	}
-	return fmt.Errorf("prooflist query %q does not match ordered traversal path %q", want, got)
 }
 
 type proofListVerifiedPath struct {
@@ -438,122 +62,9 @@ func (p proofListVerifiedPath) logicalQueryPath() string {
 	return strings.Join(p.parts, "/")
 }
 
-func (s *Server) verifyProofListStep(g graph.Runtime, index int, step prooflist.Step) (bool, error) {
-	switch step.EvidenceKind {
-	case "explicit", "implicit", "hamt":
-		ev, err := decodeEvidence(step.EvidenceKind, step.Evidence)
-		if err != nil {
-			return false, fmt.Errorf("invalid evidence at step %d: %w", index, err)
-		}
-		return g.Resolver().VerifyTranscript(step.From, &resolver.Transcript{Steps: []resolver.StepEvidence{{
-			Path:     arcset.CanonicalizePath(step.Path),
-			Target:   step.Target,
-			Evidence: ev,
-		}}})
-	case "structure":
-		switch step.EvidenceBackend {
-		case "map":
-			key := arcset.CanonicalizePath(step.Path)
-			return g.Semantic().Verify(step.From, key, mapping.Binding{Value: step.Target, Present: true}, structure.Proof(step.Proof))
-		case "list":
-			if step.Index == nil {
-				return false, fmt.Errorf("prooflist step %d list index is missing", index)
-			}
-			if step.Length == nil {
-				return false, fmt.Errorf("prooflist step %d list length is missing", index)
-			}
-			return g.ListSemantic().Verify(step.From, *step.Index, list.Query{Key: step.Target, Length: *step.Length}, structure.Proof(step.Proof))
-		case "measured_list":
-			if !step.Target.Equals(step.From) {
-				return false, nil
-			}
-			if step.Start == nil {
-				return false, fmt.Errorf("prooflist step %d list range start is missing", index)
-			}
-			if step.ChildCount == nil {
-				return false, fmt.Errorf("prooflist step %d list range child count is missing", index)
-			}
-			if step.TotalSize == nil {
-				return false, fmt.Errorf("prooflist step %d list range total size is missing", index)
-			}
-			if step.ChunkSize == nil {
-				return false, fmt.Errorf("prooflist step %d list range chunk size is missing", index)
-			}
-			measured, ok := g.ListSemantic().(list.MeasuredSemantics)
-			if !ok {
-				return false, fmt.Errorf("prooflist step %d has measured list evidence but graph list semantic does not support measured ranges", index)
-			}
-			return measured.VerifyRange(step.From, *step.Start, step.End, list.RangeResult{
-				Metadata: list.RangeMetadata{
-					ChildCount: *step.ChildCount,
-					TotalSize:  *step.TotalSize,
-					ChunkSize:  *step.ChunkSize,
-				},
-				Segments: append([]cid.Cid(nil), step.Segments...),
-			}, structure.Proof(step.Proof))
-		default:
-			return false, fmt.Errorf("prooflist step %d has unsupported structure evidence backend %q", index, step.EvidenceBackend)
-		}
-	default:
-		return false, fmt.Errorf("prooflist step %d has unsupported evidence labels %q/%q", index, step.EvidenceKind, step.EvidenceBackend)
-	}
-}
-
 var (
 	errPathNotFound = errors.New("path not found")
 )
-
-func (s *Server) resolvePath(ctx context.Context, g graph.Runtime, root cid.Cid, rawPath string, wantProof bool) (*pathResolution, error) {
-	cleanPath := querypath.CanonicalizeQueryPath(rawPath)
-	keyResult, err := g.Resolver().ResolveKey(root, cleanPath)
-	if err != nil {
-		return nil, err
-	}
-	if resolveMiss(root, cleanPath, keyResult) {
-		return nil, errPathNotFound
-	}
-	key := keyResult.Target
-	if !key.Defined() {
-		return nil, errPathNotFound
-	}
-
-	stat, target, err := s.statForResolvedKey(ctx, g, key)
-	if err != nil {
-		return nil, err
-	}
-
-	resolved := &pathResolution{
-		queryPath: cleanPath,
-		target:    target,
-		stat:      stat,
-	}
-	if !wantProof {
-		return resolved, nil
-	}
-
-	transcript := keyResult.Transcript
-	if codec.SemanticKindOf(key) == codec.SemanticKindMap {
-		payloadResult, err := g.Resolver().ResolveKey(key, explicit.PayloadArc.String())
-		if err != nil {
-			return nil, err
-		}
-		if resolveMiss(key, explicit.PayloadArc.String(), payloadResult) {
-			return nil, errPathNotFound
-		}
-		if !payloadResult.Target.Equals(target) {
-			return nil, fmt.Errorf("%w: resolved target %s does not match projected target %s", resolver.ErrResolutionFailed, payloadResult.Target, target)
-		}
-		transcript.Steps = append(transcript.Steps, payloadResult.Transcript.Steps...)
-	}
-
-	pl, err := resolver.ProofListFromTranscript(root, transcript)
-	if err != nil {
-		return nil, err
-	}
-	pl.Query = cleanPath
-	resolved.proofList = pl
-	return resolved, nil
-}
 
 func (s *Server) statForResolvedKey(ctx context.Context, g graph.Runtime, key cid.Cid) (*httpapi.PathStatResponse, cid.Cid, error) {
 	switch codec.SemanticKindOf(key) {
@@ -831,7 +342,7 @@ func (s *Server) applyUnixFSLayoutMutation(ctx context.Context, g graph.Runtime,
 }
 
 func (s *Server) applyWriterMutation(ctx context.Context, g graph.Runtime, mut writer.SemanticMutation) (writer.WriteReceipt, error) {
-	return g.Writer().Apply(ctx, g.Namespace(), mut)
+	return graphService{runtime: g}.ApplyMutation(ctx, mut)
 }
 
 func semanticMutationFromUnixFSPlan(plan *unixfs.MutationPlan, fallbackRoot cid.Cid) writer.SemanticMutation {
@@ -962,21 +473,6 @@ func (s *Server) readStatFile(ctx context.Context, g graph.Runtime, stat *httpap
 		return s.readListRange(ctx, g, key, 0, size)
 	default:
 		return nil, fmt.Errorf("unsupported file storage kind %q", stat.StorageKind)
-	}
-}
-
-func (s *Server) readContentPayload(ctx context.Context, g graph.Runtime, stat *httpapi.PathStatResponse, key cid.Cid, start, endExclusive int64) ([]byte, error) {
-	switch stat.StorageKind {
-	case "raw":
-		raw, err := s.node.CAS().Get(ctx, key)
-		if err != nil {
-			return nil, err
-		}
-		return raw[start:endExclusive], nil
-	case "list":
-		return s.readListRange(ctx, g, key, start, endExclusive)
-	default:
-		return nil, fmt.Errorf("unsupported storage kind for content")
 	}
 }
